purchase-microservice/requirements.txt
flask
flask_sqlalchemy
pymysql
requests
tenacity
cryptography